flickr8k_backend/requirements.txt
tensorflow==2.10.0
keras==2.10.0
numpy>=1.21,<1.24
pillow>=9.0.0,<10.0.0
django>=4.1,<5.0
nltk>=3.7,<4.0
scikit-learn>=1.1,<1.3
